Final implementation (100 autolab with comments)
It has autograding 100% in autolab, and everything works correct. The code has comments and good view. There are folders with tests and for output json files and output dots files. The codes are in the folder src and the zip is all the .java files that i send to autolab.

Sortly, the HW2 is to implement a program that has input a json file to create a tree, and then it can run the MinMax algorithm or the MinMax algorithm with pruning and show the optimal path for this tree. Also it evalueates all nodes of tree and generally solves the algorithm and can print JSON file of new tree, or DOT file for grapiz suite to have an optical view of tree.